bedelia/requirements.txt
# --- Core ---
Django>=4.2,<6.0
djangorestframework>=3.15,<4.0

# Filtering (DRF integration)
django-filter>=24.2,<25.0

# API docs (OpenAPI/Swagger via Spectacular)
drf-spectacular>=0.27,<1.0

# CORS for browser clients
django-cors-headers>=4.4,<5.0

# Database (PostgreSQL)
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# Static files in production
whitenoise>=6.7,<7.0

# Auth (JWT for DRF)
djangorestframework-simplejwt>=5.3,<6.0

# Production WSGI server (skip on Windows)
gunicorn>=22.0,<24.0 ; platform_system != "Windows"

python-dotenv==0.10.1